renamed files to report_writer
